upload_data/requirements.txt
azure-storage-blob
apscheduler
pytz
python-dotenv
pandas
moviepy 